moved single test case to send_and_balance.go
